requirements.txt
Flask
rdkit-pypi
